requirements.txt
toolhouse 
anthropic 
python-dotenv


